Create ScreenBuffer Class
In order to manipulate one pixel at the screen I needed
a surface. I wanted use textures but it was not as easy as
surfaces when it comes to manipulating a specific pixel.
In the future I will create textures from this surface and
use it. So It can use GPU instead of the CPU.