Create Screen Class
Screen class is needed for rendering manipulated pixel surface
to the window. It produces a game window so I can test what is On
the screen or not. Main porpose of this class will be drawing and altering
pixels at the surface and create textures from these surfaces and render
the textures. I might change this class name to PixelScreen later on.